docs(validator): sync projected README tree with builtin registries
