+++ calculogic-validator/src/core/config/validator-config.contracts.mjs
@@ -1,0 +1,1 @@
+export const VALIDATOR_CONFIG_VERSION = '0.1';
