+++ calculogic-validator/src/validator-root-files.knowledge.mjs
@@ -1,12 +1,1 @@
-export const ROOT_APP_FILES = new Set([
-  'package.json',
-  'package-lock.json',
-  'eslint.config.js',
-  'eslint.config.mjs',
-  'vite.config.ts',
-  'vite.config.js',
-  'vite.config.mjs',
-  'tsconfig.json',
-  'tsconfig.app.json',
-  'tsconfig.node.json',
-]);
+export * from './core/validator-root-files.knowledge.mjs';
