+++ calculogic-validator/test/validator-config.extensions.integration.test.mjs
@@ -4,7 +4,7 @@ import fs from 'node:fs';
 import os from 'node:os';
 import path from 'node:path';
 import { runNamingValidator } from '../src/naming/naming-validator.host.mjs';
-import { loadValidatorConfigFromFile } from '../src/validator-config.logic.mjs';
+import { loadValidatorConfigFromFile } from '../src/core/config/validator-config.logic.mjs';
 
 test('config reportableExtensions.add includes .py files in scan set', () => {
   const tempRoot = fs.mkdtempSync(path.join(os.tmpdir(), 'validator-config-ext-'));
