+++ calculogic-validator/test/validator-config.roles.integration.test.mjs
@@ -4,7 +4,7 @@ import fs from 'node:fs';
 import os from 'node:os';
 import path from 'node:path';
 import { runNamingValidator } from '../src/naming/naming-validator.host.mjs';
-import { loadValidatorConfigFromFile } from '../src/validator-config.logic.mjs';
+import { loadValidatorConfigFromFile } from '../src/core/config/validator-config.logic.mjs';
 
 test('config naming.roles.add recognizes added role and changes classification delta', () => {
   const tempRoot = fs.mkdtempSync(path.join(os.tmpdir(), 'validator-config-roles-'));
