+++ calculogic-validator/test/validator-report-meta.test.mjs
@@ -4,11 +4,11 @@ import fs from 'node:fs';
 import path from 'node:path';
 import { spawnSync } from 'node:child_process';
 import { fileURLToPath } from 'node:url';
-import { loadValidatorConfigFromFile } from '../src/validator-config.logic.mjs';
+import { loadValidatorConfigFromFile } from '../src/core/config/validator-config.logic.mjs';
 import {
   computeConfigDigest,
   getValidatorToolVersion,
-} from '../src/validator-report-meta.logic.mjs';
+} from '../src/core/validator-report-meta.logic.mjs';
 
 test('getValidatorToolVersion matches calculogic-validator/package.json version', () => {
   const packageJsonPath = path.resolve(
